Renamed main to start
Hey, I just noticed that we have a main() in both UserInterface and Game. It can't be in both. I changed this one because the signature for main was miswritten here anyway (it was "main(String args)", should have been "main(String[] args))" so we would have had to change this signature anyway.